Fix tagPrefix/vTagPrefix coherence
I thought I'd gotten the defaults right with the coherence checking, but
I never wrote a test for it so, inevitably, it wasn't..

Fixed by not giving `dynverTagPrefix` a default value so it's possible
to tell if a user set it or not.

Now this sets the tagPrefix to "" if vTagPrefix was set to false.

If both keys are set, to incoherent values, it will throw.