Fix SHA longer than 8 chars cannot be matched (#231)
In case 8 characters aren't enough to uniquely identify a commit, the
SHA will be longer. Because the regexes that match the SHA expect
exactly 8 characters, stuff breaks in this case. The fix allows for
longer SHAs, but in the case of the parser only up to 39 (40 would be
the total length). This is necessary, because otherwise tests fail,
because if there isn't a tag, then we only ever see the SHA when
requesting the previous version, and we must not match it.